% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBS.xcms_diffreport.R
\docType{data}
\name{CBS.xcms_diffreport}
\alias{CBS.xcms_diffreport}
\title{An 'xcms' diffreport object is included as limms package data}
\format{
A data frame with 274 rows and 69 columns
}
\usage{
data(CBS.xcms_diffreport)
}
\description{
Mass spectrometry peak list of Saccharomyces cerevisiae metabolites after
complementation with human cystathionine beta-synthase (CBS) alleles.
LC-MS data from an Orbitrap with peaks picked and grouped using
the R package 'xcms'.
Data includes replicate samples with two different CBS alleles,
with or without methionine, with or without vitamin B6 cofactor,
and samples spiked with differeing concentrations of pooled,
isotopically-labelled metabolites.
}
\examples{
data(CBS.xcms_diffreport)
names(CBS.xcms_diffreport)
}
\references{
\url{https://pubmed.ncbi.nlm.nih.gov/22267502/}
}
\author{
Jacob Mayfield
}
\keyword{data}
