% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmdb_Shortlist.R
\docType{data}
\name{hmdb_Shortlist}
\alias{hmdb_Shortlist}
\title{Metabolites included as calabration standards in the Saccharomyces CBS data}
\format{
A data frame with 27 rows and 5 columns
}
\usage{
hmdb_Shortlist
}
\description{
An database of known metabolites included as limms package data
HMDB_Shortlist contains a list of compounds 
used as calibration standards in the CBS experiment.
}
\examples{
data(hmdb_Shortlist)
names(hmdb_Shortlist)
}
\references{
\url{https://pubmed.ncbi.nlm.nih.gov/22267502/}
}
\author{
Jacob Mayfield
}
\keyword{data}
