% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmCalc.R
\name{ppmCalc}
\alias{ppmCalc}
\title{Calculate parts per million (ppm)}
\usage{
ppmCalc(expected, observed)
}
\arguments{
\item{expected}{The expected m/z value.}

\item{observed}{The measured m/z value.}
}
\value{
Returns the deviation from the expected m/z in parts per million.
}
\description{
This function calculates m/z deviations in parts per millon (ppm).
}
\details{
A simple ppm calculator and internal helper function in limms
}
\examples{
# Is a [M+H]+ measured at 90.0552 alanine?  

ppmCalc(90.0550, 90.0552)
# May not be alanine, but with a ppm < 2.23 its close.
}
\seealso{
\code{\link{dbMatch}}
}
