% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttF.CBS.R
\docType{data}
\name{ttF.CBS}
\alias{ttF.CBS}
\title{'limma' topTable object of CBS metabolomic data}
\format{
A data frame with 52 rows and 9 columns
}
\usage{
ttF.CBS
}
\description{
The metabolite data in the data set 'CBS.xcms_diffreport'
analyzed using the limms pipeline to generate a table of
adjusted P values based on an F-score after contrasts of
all different sample classes.
The dataset is the output of the 'limma' function 'topTable',
available in the 'limms' function 'limmaTest'
}
\examples{
data(ttF.CBS)
}
\references{
\url{https://pubmed.ncbi.nlm.nih.gov/22267502/}
}
\seealso{
\code{\link{CBS.xcms_diffreport}}
\code{\link{limmaTest}}
}
\author{
Jacob Mayfield
}
\keyword{data}
